+++ src/storage/external_sort.c
@@ -1484,10 +1484,10 @@ sort_listfile (THREAD_ENTRY * thread_p, INT16 volid, int est_inp_pg_cnt, SORT_GE
        * NEED MORE CONSIDERATION */
       num_cpus = fileio_os_sysconf ();
 
-      sort_param->px_height_max = (int) sqrt (num_cpus);	/* n */
+      sort_param->px_height_max = (int) sqrt ((double) num_cpus);	/* n */
       sort_param->px_array_size = num_cpus;	/* 2^^n */
 
-      assert_release (sort_param->px_array_size == pow (2, sort_param->px_height_max));
+      assert_release (sort_param->px_array_size == pow ((double) 2, (double)  sort_param->px_height_max));
     }
 #endif /* SERVER_MODE */
